Remove expect/actual pattern for ImpressumSection and ImpressumWrapper
Move impressum UI components to commonMain with isPlatformWasm checks.
Only use expect/actual for platform-specific URL opening functionality.

- Deleted platform-specific ImpressumSection implementations
- Deleted Impressum.wasmJs.kt
- Moved ImpressumWrapper and ImpressumSection to commonMain
- Added UrlOpener expect/actual for browser link opening
- All platforms compile successfully
