--- composeApp/src/androidMain/kotlin/de/egril/defender/ui/infopage/ImpressumSection.android.kt
@@ -1,20 +1,0 @@
-package de.egril.defender.ui.infopage
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-/**
- * Impressum section for the info page - Android implementation (not shown)
- */
-@Composable
-actual fun ImpressumSection() {
-    // No impressum on Android platform
-}
-
-/**
- * Impressum wrapper - Android implementation (not shown)
- */
-@Composable
-actual fun ImpressumWrapper(rowModifier: Modifier) {
-    // No impressum on Android platform
-}
