+++ composeApp/src/androidMain/kotlin/de/egril/defender/ui/infopage/UrlOpener.android.kt
@@ -1,0 +1,8 @@
+package de.egril.defender.ui.infopage
+
+/**
+ * Opens a URL in the browser - Android implementation (no-op).
+ */
+actual fun openUrl(url: String) {
+    // No-op on Android - impressum won't be shown anyway
+}
