+++ composeApp/src/commonMain/kotlin/de/egril/defender/ui/infopage/UrlOpener.kt
@@ -1,0 +1,7 @@
+package de.egril.defender.ui.infopage
+
+/**
+ * Opens a URL in the browser.
+ * This is implemented platform-specifically.
+ */
+expect fun openUrl(url: String)
