--- composeApp/src/desktopMain/kotlin/de/egril/defender/ui/infopage/ImpressumSection.desktop.kt
@@ -1,20 +1,0 @@
-package de.egril.defender.ui.infopage
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-/**
- * Impressum section for the info page - Desktop implementation (not shown)
- */
-@Composable
-actual fun ImpressumSection() {
-    // No impressum on desktop platform
-}
-
-/**
- * Impressum wrapper - Desktop implementation (not shown)
- */
-@Composable
-actual fun ImpressumWrapper(rowModifier: Modifier) {
-    // No impressum on desktop platform
-}
