--- composeApp/src/iosMain/kotlin/de/egril/defender/ui/infopage/ImpressumSection.ios.kt
@@ -1,20 +1,0 @@
-package de.egril.defender.ui.infopage
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-/**
- * Impressum section for the info page - iOS implementation (not shown)
- */
-@Composable
-actual fun ImpressumSection() {
-    // No impressum on iOS platform
-}
-
-/**
- * Impressum wrapper - iOS implementation (not shown)
- */
-@Composable
-actual fun ImpressumWrapper(rowModifier: Modifier) {
-    // No impressum on iOS platform
-}
