+++ composeApp/src/wasmJsMain/kotlin/de/egril/defender/ui/infopage/UrlOpener.wasmJs.kt
@@ -1,0 +1,8 @@
+package de.egril.defender.ui.infopage
+
+/**
+ * Opens a URL in the browser - WASM implementation.
+ */
+actual fun openUrl(url: String) {
+    kotlinx.browser.window.open(url, "_blank")
+}
